drop support for multiple tokens in NFT contract
There is no reasons why we should support multiple tokens at mint time,
it would only make things difficult to understand when it comes to raffle prize and such.

The support for multiple tokens is there for the Raffler to support listing on
NFT marketplaces and support the tokens used on those markets.